Refactor procedural code into objects. Remove use of arguments.
The aim of the refactor is to improve readability, ease of debugging,
and open the possibility for tests.

Use of command-line arguments has been removed because they're not
secure (they'll appear in your shell history, for example). Instead,
using Foreman and an .env file is recommended.

Drop support for Ruby 1.8.7.

Add example .env file.

Update README.